# Conflicts:
#	public/app/plugins/datasource/azuremonitor/components/AzureCredentialsForm.tsx

+++ pkg/services/ssosettings/ssosettingsimpl/service.go
@@ -2,6 +2,7 @@ package ssosettingsimpl
 
 import (
 	"context"
+	"encoding/base64"
 	"errors"
 	"fmt"
 	"strings"
@@ -24,7 +25,7 @@ import (
 var _ ssosettings.Service = (*SSOSettingsService)(nil)
 
 type SSOSettingsService struct {
-	log     log.Logger
+	logger  log.Logger
 	cfg     *setting.Cfg
 	store   ssosettings.Store
 	ac      ac.AccessControl
@@ -45,7 +46,7 @@ func ProvideService(cfg *setting.Cfg, sqlStore db.DB, ac ac.AccessControl,
 	store := database.ProvideStore(sqlStore)
 
 	svc := &SSOSettingsService{
-		log:          log.New("ssosettings.service"),
+		logger:       log.New("ssosettings.service"),
 		cfg:          cfg,
 		store:        store,
 		ac:           ac,
@@ -65,35 +66,40 @@ func ProvideService(cfg *setting.Cfg, sqlStore db.DB, ac ac.AccessControl,
 var _ ssosettings.Service = (*SSOSettingsService)(nil)
 
 func (s *SSOSettingsService) GetForProvider(ctx context.Context, provider string) (*models.SSOSettings, error) {
-	storeSettings, err := s.store.Get(ctx, provider)
+	dbSettings, err := s.store.Get(ctx, provider)
+	if err != nil && !errors.Is(err, ssosettings.ErrNotFound) {
+		return nil, err
+	}
 
-	if errors.Is(err, ssosettings.ErrNotFound) {
-		settings, err := s.loadSettingsUsingFallbackStrategy(ctx, provider)
+	if dbSettings != nil {
+		// Settings are coming from the database thus secrets are encrypted
+		dbSettings.Settings, err = s.decryptSecrets(ctx, dbSettings.Settings)
 		if err != nil {
 			return nil, err
 		}
-
-		return settings, nil
 	}
 
+	systemSettings, err := s.loadSettingsUsingFallbackStrategy(ctx, provider)
 	if err != nil {
 		return nil, err
 	}
 
-	storeSettings.Source = models.DB
-
-	return storeSettings, nil
+	return s.mergeSSOSettings(dbSettings, systemSettings), nil
 }
 
 func (s *SSOSettingsService) GetForProviderWithRedactedSecrets(ctx context.Context, provider string) (*models.SSOSettings, error) {
+	if !s.isProviderConfigurable(provider) {
+		return nil, ssosettings.ErrNotConfigurable
+	}
+
 	storeSettings, err := s.GetForProvider(ctx, provider)
 	if err != nil {
 		return nil, err
 	}
 
 	for k, v := range storeSettings.Settings {
-		if isSecret(k) && v != "" {
-			storeSettings.Settings[k] = "*********"
+		if strVal, ok := v.(string); ok {
+			storeSettings.Settings[k] = setting.RedactedValue(k, strVal)
 		}
 	}
 
@@ -109,25 +115,52 @@ func (s *SSOSettingsService) List(ctx context.Context) ([]*models.SSOSettings, e
 	}
 
 	for _, provider := range ssosettings.AllOAuthProviders {
-		settings := getSettingsByProvider(provider, storedSettings)
-		if len(settings) == 0 {
-			// If there is no data in the DB then we need to load the settings using the fallback strategy
-			fallbackSettings, err := s.loadSettingsUsingFallbackStrategy(ctx, provider)
+		dbSettings := getSettingByProvider(provider, storedSettings)
+		if dbSettings != nil {
+			// Settings are coming from the database thus secrets are encrypted
+			dbSettings.Settings, err = s.decryptSecrets(ctx, dbSettings.Settings)
 			if err != nil {
 				return nil, err
 			}
-
-			settings = append(settings, fallbackSettings)
 		}
-		result = append(result, settings...)
+		fallbackSettings, err := s.loadSettingsUsingFallbackStrategy(ctx, provider)
+		if err != nil {
+			return nil, err
+		}
+
+		result = append(result, s.mergeSSOSettings(dbSettings, fallbackSettings))
 	}
 
 	return result, nil
 }
 
-func (s *SSOSettingsService) Upsert(ctx context.Context, settings models.SSOSettings) error {
-	if !isProviderConfigurable(settings.Provider) {
-		return ssosettings.ErrInvalidProvider.Errorf("provider %s is not configurable", settings.Provider)
+func (s *SSOSettingsService) ListWithRedactedSecrets(ctx context.Context) ([]*models.SSOSettings, error) {
+	storeSettings, err := s.List(ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	configurableSettings := make([]*models.SSOSettings, 0, len(s.cfg.SSOSettingsConfigurableProviders))
+	for _, provider := range storeSettings {
+		if s.isProviderConfigurable(provider.Provider) {
+			configurableSettings = append(configurableSettings, provider)
+		}
+	}
+
+	for _, storeSetting := range configurableSettings {
+		for k, v := range storeSetting.Settings {
+			if strVal, ok := v.(string); ok {
+				storeSetting.Settings[k] = setting.RedactedValue(k, strVal)
+			}
+		}
+	}
+
+	return configurableSettings, nil
+}
+
+func (s *SSOSettingsService) Upsert(ctx context.Context, settings *models.SSOSettings) error {
+	if !s.isProviderConfigurable(settings.Provider) {
+		return ssosettings.ErrNotConfigurable
 	}
 
 	social, ok := s.reloadables[settings.Provider]
@@ -135,21 +168,19 @@ func (s *SSOSettingsService) Upsert(ctx context.Context, settings models.SSOSett
 		return ssosettings.ErrInvalidProvider.Errorf("provider %s not found in reloadables", settings.Provider)
 	}
 
-	err := social.Validate(ctx, settings)
+	err := social.Validate(ctx, *settings)
 	if err != nil {
 		return err
 	}
 
-	systemSettings, err := s.loadSettingsUsingFallbackStrategy(ctx, settings.Provider)
+	storedSettings, err := s.GetForProvider(ctx, settings.Provider)
 	if err != nil {
 		return err
 	}
 
-	// add the SSO settings from system that are not available in the user input
-	// in order to have a complete set of SSO settings for every provider in the database
-	settings.Settings = mergeSettings(settings.Settings, systemSettings.Settings)
+	secrets := collectSecrets(settings, storedSettings)
 
-	settings.Settings, err = s.encryptSecrets(ctx, settings.Settings)
+	settings.Settings, err = s.encryptSecrets(ctx, settings.Settings, storedSettings.Settings)
 	if err != nil {
 		return err
 	}
@@ -160,9 +191,10 @@ func (s *SSOSettingsService) Upsert(ctx context.Context, settings models.SSOSett
 	}
 
 	go func() {
-		err = social.Reload(context.Background(), settings)
+		settings.Settings = overrideMaps(storedSettings.Settings, settings.Settings, secrets)
+		err = social.Reload(context.Background(), *settings)
 		if err != nil {
-			s.log.Error("failed to reload the provider", "provider", settings.Provider, "error", err)
+			s.logger.Error("failed to reload the provider", "provider", settings.Provider, "error", err)
 		}
 	}()
 
@@ -174,6 +206,9 @@ func (s *SSOSettingsService) Patch(ctx context.Context, provider string, data ma
 }
 
 func (s *SSOSettingsService) Delete(ctx context.Context, provider string) error {
+	if !s.isProviderConfigurable(provider) {
+		return ssosettings.ErrNotConfigurable
+	}
 	return s.store.Delete(ctx, provider)
 }
 
@@ -210,14 +245,13 @@ func (s *SSOSettingsService) loadSettingsUsingFallbackStrategy(ctx context.Conte
 	}, nil
 }
 
-func getSettingsByProvider(provider string, settings []*models.SSOSettings) []*models.SSOSettings {
-	result := make([]*models.SSOSettings, 0)
+func getSettingByProvider(provider string, settings []*models.SSOSettings) *models.SSOSettings {
 	for _, item := range settings {
 		if item.Provider == provider {
-			result = append(result, item)
+			return item
 		}
 	}
-	return result
+	return nil
 }
 
 func (s *SSOSettingsService) getFallbackStrategyFor(provider string) (ssosettings.FallbackStrategy, bool) {
@@ -229,20 +263,24 @@ func (s *SSOSettingsService) getFallbackStrategyFor(provider string) (ssosetting
 	return nil, false
 }
 
-func (s *SSOSettingsService) encryptSecrets(ctx context.Context, settings map[string]any) (map[string]any, error) {
+func (s *SSOSettingsService) encryptSecrets(ctx context.Context, settings, storedSettings map[string]any) (map[string]any, error) {
 	result := make(map[string]any)
 	for k, v := range settings {
-		if isSecret(k) {
+		if isSecret(k) && v != "" {
 			strValue, ok := v.(string)
 			if !ok {
 				return result, fmt.Errorf("failed to encrypt %s setting because it is not a string: %v", k, v)
+			}
+
+			if !isNewSecretValue(strValue) {
+				strValue = storedSettings[k].(string)
 			}
 
 			encryptedSecret, err := s.secrets.Encrypt(ctx, []byte(strValue), secrets.WithoutScope())
 			if err != nil {
 				return result, err
 			}
-			result[k] = string(encryptedSecret)
+			result[k] = base64.RawStdEncoding.EncodeToString(encryptedSecret)
 		} else {
 			result[k] = v
 		}
@@ -252,7 +290,12 @@ func (s *SSOSettingsService) encryptSecrets(ctx context.Context, settings map[st
 }
 
 func (s *SSOSettingsService) Run(ctx context.Context) error {
-	ticker := time.NewTicker(1 * time.Minute)
+	interval := s.cfg.SSOSettingsReloadInterval
+	if interval == 0 {
+		return nil
+	}
+
+	ticker := time.NewTicker(interval)
 
 	// start a background process for reloading the SSO settings for all providers at a fixed interval
 	// it is useful for high availability setups running multiple Grafana instances
@@ -268,25 +311,135 @@ func (s *SSOSettingsService) Run(ctx context.Context) error {
 }
 
 func (s *SSOSettingsService) doReload(ctx context.Context) {
-	s.log.Debug("reloading SSO Settings for all providers")
+	s.logger.Debug("reloading SSO Settings for all providers")
 
 	settingsList, err := s.List(ctx)
 	if err != nil {
-		s.log.Error("failed to fetch SSO Settings for all providers", "err", err)
+		s.logger.Error("failed to fetch SSO Settings for all providers", "err", err)
 		return
 	}
 
 	for provider, connector := range s.reloadables {
-		settings := getSettingsByProvider(provider, settingsList)
+		setting := getSettingByProvider(provider, settingsList)
 
-		if len(settings) > 0 {
-			err = connector.Reload(ctx, *settings[0])
-			if err != nil {
-				s.log.Error("failed to reload SSO Settings", "provider", provider, "err", err)
-				continue
-			}
+		err = connector.Reload(ctx, *setting)
+		if err != nil {
+			s.logger.Error("failed to reload SSO Settings", "provider", provider, "err", err)
+			continue
 		}
 	}
+}
+
+// mergeSSOSettings merges the settings from the database with the system settings
+// Required because it is possible that the user has configured some of the settings (current Advanced OAuth settings)
+// and the rest of the settings have to be loaded from the system settings
+func (s *SSOSettingsService) mergeSSOSettings(dbSettings, systemSettings *models.SSOSettings) *models.SSOSettings {
+	if dbSettings == nil {
+		s.logger.Debug("No SSO Settings found in the database, using system settings")
+		return systemSettings
+	}
+
+	s.logger.Debug("Merging SSO Settings", "dbSettings", removeSecrets(dbSettings.Settings), "systemSettings", removeSecrets(systemSettings.Settings))
+
+	result := &models.SSOSettings{
+		Provider: dbSettings.Provider,
+		Source:   dbSettings.Source,
+		Settings: mergeSettings(dbSettings.Settings, systemSettings.Settings),
+		Created:  dbSettings.Created,
+		Updated:  dbSettings.Updated,
+	}
+
+	return result
+}
+
+func (s *SSOSettingsService) decryptSecrets(ctx context.Context, settings map[string]any) (map[string]any, error) {
+	for k, v := range settings {
+		if isSecret(k) && v != "" {
+			strValue, ok := v.(string)
+			if !ok {
+				s.logger.Error("Failed to parse secret value, it is not a string", "key", k)
+				return nil, fmt.Errorf("secret value is not a string")
+			}
+
+			decoded, err := base64.RawStdEncoding.DecodeString(strValue)
+			if err != nil {
+				s.logger.Error("Failed to decode secret string", "err", err, "value")
+				return nil, err
+			}
+
+			decrypted, err := s.secrets.Decrypt(ctx, decoded)
+			if err != nil {
+				s.logger.Error("Failed to decrypt secret", "err", err)
+				return nil, err
+			}
+
+			settings[k] = string(decrypted)
+		}
+	}
+	return settings, nil
+}
+
+func (s *SSOSettingsService) isProviderConfigurable(provider string) bool {
+	_, ok := s.cfg.SSOSettingsConfigurableProviders[provider]
+	return ok
+}
+
+// removeSecrets removes all the secrets from the map and replaces them with a redacted password
+// and returns a new map
+func removeSecrets(settings map[string]any) map[string]any {
+	result := make(map[string]any)
+	for k, v := range settings {
+		if isSecret(k) {
+			result[k] = setting.RedactedPassword
+			continue
+		}
+		result[k] = v
+	}
+	return result
+}
+
+// mergeSettings merges two maps in a way that the values from the first map are preserved
+// and the values from the second map are added only if they don't exist in the first map
+func mergeSettings(storedSettings, systemSettings map[string]any) map[string]any {
+	settings := make(map[string]any)
+
+	for k, v := range storedSettings {
+		settings[k] = v
+	}
+
+	for k, v := range systemSettings {
+		if _, ok := settings[k]; !ok {
+			settings[k] = v
+		}
+	}
+
+	return settings
+}
+
+// collectSecrets collects all the secrets from the request and the currently stored settings
+// and returns a new map
+func collectSecrets(settings *models.SSOSettings, storedSettings *models.SSOSettings) map[string]any {
+	secrets := map[string]any{}
+	for k, v := range settings.Settings {
+		if isSecret(k) {
+			if isNewSecretValue(v.(string)) {
+				secrets[k] = v.(string) // use the new value
+				continue
+			}
+			secrets[k] = storedSettings.Settings[k] // keep the currently stored value
+		}
+	}
+	return secrets
+}
+
+func overrideMaps(maps ...map[string]any) map[string]any {
+	result := make(map[string]any)
+	for _, m := range maps {
+		for k, v := range m {
+			result[k] = v
+		}
+	}
+	return result
 }
 
 func isSecret(fieldName string) bool {
@@ -300,28 +453,6 @@ func isSecret(fieldName string) bool {
 	return false
 }
 
-func mergeSettings(apiSettings, systemSettings map[string]any) map[string]any {
-	settings := make(map[string]any)
-
-	for k, v := range apiSettings {
-		settings[k] = v
-	}
-
-	for k, v := range systemSettings {
-		if _, ok := settings[k]; !ok {
-			settings[k] = v
-		}
-	}
-
-	return settings
-}
-
-func isProviderConfigurable(provider string) bool {
-	for _, configurable := range ssosettings.ConfigurableOAuthProviders {
-		if provider == configurable {
-			return true
-		}
-	}
-
-	return false
+func isNewSecretValue(value string) bool {
+	return value != setting.RedactedPassword
 }
